Move extend_batch function
