+++ tests/test_models.py
@@ -22,15 +22,7 @@ import pykeen.experiments
 import pykeen.models
 from pykeen.datasets.kinships import KINSHIPS_TRAIN_PATH
 from pykeen.datasets.nations import NATIONS_TEST_PATH, NATIONS_TRAIN_PATH, Nations
-from pykeen.models import _MODELS
-from pykeen.models.base import (
-    EntityEmbeddingModel,
-    EntityRelationEmbeddingModel,
-    Model,
-    MultimodalModel,
-    OModel,
-    _extend_batch,
-)
+from pykeen.models import EntityEmbeddingModel, EntityRelationEmbeddingModel, Model, MultimodalModel, OModel, _MODELS
 from pykeen.models.cli import build_cli_from_cls
 from pykeen.models.predict import get_novelty_mask, predict
 from pykeen.models.unimodal.rgcn import (
@@ -42,7 +34,7 @@ from pykeen.models.unimodal.trans_d import _project_entity
 from pykeen.nn import Embedding, RepresentationModule
 from pykeen.training import LCWATrainingLoop, SLCWATrainingLoop, TrainingLoop
 from pykeen.triples import TriplesFactory
-from pykeen.utils import all_in_bounds, clamp_norm, set_random_seed
+from pykeen.utils import all_in_bounds, clamp_norm, extend_batch, set_random_seed
 
 SKIP_MODULES = {
     Model.__name__,
@@ -1276,7 +1268,7 @@ class TestModelUtilities(unittest.TestCase):
         num_choices = len(all_ids)
 
         for dim in range(3):
-            h_ext_batch = _extend_batch(batch=batch, all_ids=all_ids, dim=dim)
+            h_ext_batch = extend_batch(batch=batch, all_ids=all_ids, dim=dim)
 
             # check shape
             assert h_ext_batch.shape == (batch_size * num_choices, 3)
